recommendation/requirements.txt
greenlet==3.0.0
grpcio==1.59.0
grpcio-tools==1.59.0
joblib==1.3.2
numpy==1.26.1
pandas==2.1.1
protobuf==4.24.4
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
SQLAlchemy==2.0.22
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
python-dotenv==1.0.0